+++ zotify/app.py
@@ -149,28 +149,27 @@ class App:
         Logger(self.__config)
 
         # Create session
-        # if args.username != "" and args.password != "":
-        #     self.__session = Session.from_userpass(
-        #         args.username,
-        #         args.password,
-        #         self.__config.credentials_path,
-        #         self.__config.language,
-        #     )
-        # elif self.__config.credentials_path.is_file():
-        #     self.__session = Session.from_file(
-        #         self.__config.credentials_path, self.__config.language
-        #     )
-        # else:
-        #     self.__session = Session.from_prompt(
-        #         self.__config.credentials_path, self.__config.language
-        #     )
-        username = input("Username: ")
-        auth = OAuth(username)
-        auth_url = auth.get_authorization_url()
-        print(f"\nClick on the following link to login:\n{auth_url}")
-        self.__session = Session.from_oauth(
-            auth, self.__config.credentials_path, self.__config.language
-        )
+        if args.username != "" and args.token != "":
+            oauth = OAuth(args.username)
+            oauth.set_token(args.token, OAuth.RequestType.REFRESH)
+            self.__session = Session.from_oauth(
+                oauth, self.__config.credentials_path, self.__config.language
+            )
+        elif self.__config.credentials_path.is_file():
+            self.__session = Session.from_file(
+                self.__config.credentials_path,
+                self.__config.language,
+            )
+        else:
+            username = args.username
+            while username == "":
+                username = input("Username: ")
+            oauth = OAuth(username)
+            auth_url = oauth.auth_interactive()
+            print(f"\nClick on the following link to login:\n{auth_url}")
+            self.__session = Session.from_oauth(
+                oauth, self.__config.credentials_path, self.__config.language
+            )
 
         # Get items to download
         ids = self.get_selection(args)